+++ apps/documentation/src/components/sanity/text/LinkResolver.tsx
@@ -11,8 +11,29 @@ type Props = {
 };
 
 export const LinkResolver = ({ value }: Props) => {
-  const { linkText, linkAddress, linkType, iconName } = value;
+  const {
+    linkText,
+    linkAddress,
+    linkType,
+    linkAddressType,
+    iconName,
+    downloadFile,
+  } = value;
   const Icon = getIconByName(iconName);
+
+  // Determine the href based on linkAddressType
+  const href =
+    linkAddressType === 'file' && downloadFile?.asset?.url
+      ? downloadFile.asset.url + '?&dl='
+      : linkAddress;
+
+  const linkProps = {
+    href,
+    download:
+      linkAddressType === 'file'
+        ? downloadFile?.asset?.originalFilename
+        : undefined,
+  };
 
   switch (linkType) {
     case 'navigationcard':
@@ -20,28 +41,48 @@ export const LinkResolver = ({ value }: Props) => {
         <NavigationCard
           compact
           title={linkText ?? ''}
-          href={linkAddress}
           titleIcon={Icon ? <Icon /> : undefined}
           style={{ maxWidth: '22.5rem' }}
+          {...linkProps}
         ></NavigationCard>
       );
     case 'button':
       return (
-        <PrimaryButton as="a" href={linkAddress}>
+        <PrimaryButton
+          as="a"
+          {...linkProps}
+          style={{ marginBlockEnd: '2rem ' }}
+        >
           {linkText}
           {Icon ? <Icon /> : undefined}
         </PrimaryButton>
       );
     case 'button-secondary':
       return (
-        <SecondaryButton as="a" href={linkAddress}>
+        <SecondaryButton
+          as="a"
+          {...linkProps}
+          style={{ marginBlockEnd: '2rem ' }}
+        >
+          {linkText}
+          {Icon ? <Icon /> : undefined}
+        </SecondaryButton>
+      );
+    case 'button-secondary-small':
+      return (
+        <SecondaryButton
+          as="a"
+          size="small"
+          {...linkProps}
+          style={{ marginBlockEnd: '2rem ' }}
+        >
           {linkText}
           {Icon ? <Icon /> : undefined}
         </SecondaryButton>
       );
     default:
       return (
-        <Link href={linkAddress}>
+        <Link {...linkProps}>
           {linkText}
           {Icon ? <Icon /> : undefined}
         </Link>
@@ -56,6 +97,14 @@ export const LinkFragment = graphql`
     linkAddress
     linkText
     linkType
+    linkAddressType
     iconName
+    downloadFile {
+      asset {
+        _id
+        url
+        originalFilename
+      }
+    }
   }
 `;