+++ packages/dropdown/src/MultiSelect.tsx
@@ -222,20 +222,31 @@ export const MultiSelect = React.forwardRef(
       filterListItems({ inputValue });
     }, [normalizedItems]); // eslint-disable-line react-hooks/exhaustive-deps
 
-    const { hasSelectedItems, handleListItemClicked, selectAllCheckboxState } =
-      useMultiselectUtils<ValueType>({
-        listItems,
-        selectAll,
-        selectedItems,
-      });
+    const {
+      hasSelectedItems,
+      handleListItemClicked,
+      selectAllCheckboxState,
+      clickedItemIsInSelectedItems,
+      clickedItemIsSelectAll,
+    } = useMultiselectUtils<ValueType>({
+      listItems,
+      selectAll,
+      selectedItems,
+    });
 
-    const { getSelectedItemProps, getDropdownProps } = useMultipleSelection({
+    const {
+      getSelectedItemProps,
+      getDropdownProps,
+      reset,
+      removeSelectedItem,
+      setSelectedItems,
+    } = useMultipleSelection({
       selectedItems,
       // @ts-expect-error prop missing from library types
       itemToString,
       itemToKey,
-      onStateChange({ selectedItems: newSelectedItems }) {
-        if (newSelectedItems !== undefined) onChange(newSelectedItems);
+      onSelectedItemsChange({ selectedItems: newSelectedItems }) {
+        onChange(newSelectedItems);
       },
     });
 
@@ -259,7 +270,6 @@ export const MultiSelect = React.forwardRef(
         switch (type) {
           // reset input value when leaving input field
           case useCombobox.stateChangeTypes.InputBlur:
-            if (state.isOpen) inputRef.current?.focus();
             return {
               ...changes,
               inputValue: EMPTY_INPUT,
@@ -341,13 +351,13 @@ export const MultiSelect = React.forwardRef(
         setHighlightedIndex(hideSelectAll ? 0 : 1);
         setLastHighlightedIndex(hideSelectAll ? 0 : 1);
       },
-      onStateChange({ selectedItem: clickedItem }) {
+      onSelectedItemChange({ selectedItem: clickedItem }) {
         // clickedItem means item chosen either via mouse or keyboard
         if (!clickedItem) return;
 
         handleListItemClicked({
           clickedItem,
-          onChange,
+          onChange: setSelectedItems,
         });
       },
       // Accessibility
@@ -396,10 +406,8 @@ export const MultiSelect = React.forwardRef(
     }, [isOpen, refs.reference, refs.floating, update]);
 
     const handleOnClear = () => {
-      onChange([]);
-      setInputValue(EMPTY_INPUT);
       inputRef.current?.focus();
-      updateListItems({ inputValue });
+      reset();
     };
 
     return (
@@ -481,10 +489,7 @@ export const MultiSelect = React.forwardRef(
                 }
                 readOnly={readOnly}
                 removeSelectedItem={() => {
-                  handleListItemClicked({
-                    clickedItem: selectedItem,
-                    onChange,
-                  });
+                  removeSelectedItem(selectedItem);
                   inputRef?.current?.focus();
                 }}
                 selectedItem={selectedItem}
@@ -505,13 +510,20 @@ export const MultiSelect = React.forwardRef(
               onKeyDown: (e: React.KeyboardEvent) => {
                 if (selectOnTab && isOpen && e.key === 'Tab') {
                   const highlitedItem = listItems[highlightedIndex];
-                  // we don't want to clear selection with tab
-                  if (highlitedItem) {
-                    handleListItemClicked({
-                      clickedItem: highlitedItem,
-                      onChange,
-                    });
-                  }
+                  if (!highlitedItem) return;
+
+                  // Skip tab selection for select all or if item already is selected
+                  const shouldSkipTabSelection =
+                    clickedItemIsSelectAll(highlitedItem) ||
+                    (!clickedItemIsSelectAll(highlitedItem) &&
+                      clickedItemIsInSelectedItems(highlitedItem));
+
+                  if (shouldSkipTabSelection) return;
+
+                  handleListItemClicked({
+                    clickedItem: highlitedItem,
+                    onChange: setSelectedItems,
+                  });
                 }
               },
               ...getDropdownProps({