+++ apps/code-playground/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.9](https://github.com/entur/design-system/compare/@entur/code-playground@2.1.8...@entur/code-playground@2.1.9) (2025-08-01)
+
+**Note:** Version bump only for package @entur/code-playground
+
+## [2.1.8](https://github.com/entur/design-system/compare/@entur/code-playground@2.1.7...@entur/code-playground@2.1.8) (2025-07-29)
+
+**Note:** Version bump only for package @entur/code-playground
+
 ## [2.1.7](https://github.com/entur/design-system/compare/@entur/code-playground@2.1.6...@entur/code-playground@2.1.7) (2025-06-27)
 
 **Note:** Version bump only for package @entur/code-playground